+++ src/app/(auth)/login/_components/LoginForm/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ForgotPasswordButton from '@/app/(auth)/login/_components/LoginForm/ForgotPasswordButton';
+import PasswordResetRequestModal from '@/app/(auth)/login/_components/LoginForm/PasswordResetRequestModal';
 import InputWithLabel from '@/components/auth/InputWithLabel';
 import Button from '@/components/common/Button';
 import { signIn } from '@/lib/apis/auth';
@@ -205,7 +205,7 @@ export default function LoginForm() {
         />
       </div>
 
-      <ForgotPasswordButton />
+      <PasswordResetRequestModal />
 
       <Button
         size="lg"
